[model/UGC] improve get_ugc_metadata_doctypes, keep backward compat
MySQL's SQL_CALC_FOUND_ROWS forces a full result set generation
voiding the benefits of the LIMIT clause. Workaround this deficency
by returning a generic "has_more" boolean instead of the actual
number of results found (number of documents available).

However, this breaks API. While we don't care too much about the
non-advertised but yet public API, the Service API is critical, since
it is used by Rigo and friends. For this reason, the new function
will be used if and only if the http request contains "revision=1"
in its parameters.